split out model mappings
