+++ bluemira/codes/process/_setup.py
@@ -27,12 +27,12 @@ from typing import ClassVar, Dict, Optional, Union
 from bluemira.codes.error import CodesError
 from bluemira.codes.interface import CodesSetup
 from bluemira.codes.process._inputs import ProcessInputs
-from bluemira.codes.process.api import InDat, _INVariable, update_obsolete_vars
-from bluemira.codes.process.constants import NAME as PROCESS_NAME
-from bluemira.codes.process.mapping import (
+from bluemira.codes.process._model_mapping import (
     CurrentDriveEfficiencyModel,
     TFCoilConductorTechnology,
 )
+from bluemira.codes.process.api import InDat, _INVariable, update_obsolete_vars
+from bluemira.codes.process.constants import NAME as PROCESS_NAME
 from bluemira.codes.process.params import ProcessSolverParams
