+++ bluemira/codes/process/api.py
@@ -26,7 +26,7 @@ from dataclasses import dataclass
 from enum import Enum
 from importlib import resources
 from pathlib import Path
-from typing import Dict, List, Optional, TypeVar, Union
+from typing import Dict, List, TypeVar, Union
 
 from bluemira.base.look_and_feel import bluemira_print, bluemira_warn
 from bluemira.codes.error import CodesError
